Player Health Bar and quick fixes to menus
Created the script for the player's health and its health bar UI element. When the player takes damages from projectiles, its health is decreased (which shows on the healthbar), and if its health reaches zero, he dies and the game goes to the Game Over scene.
Also removed Settings button from menus