Fixes to Pause Menu, creation of assets
Fixed the Pause Menu bugs, now it automatically highlights the first button in the menu, "Resume Game", can be navigated with arrow keys, and works without other bugs like it would be inconsistent with the highlighting. Also created assets for the Canvas_PauseMenu, the PauseMenu itself, its EventManager, the Canvas_HealthBar, and the Healthbar.

FINALLY, my build doesn't have the completed boss, and in its current state it keeps pausing the scene (not like opening the Pause Menu, pausing as in the button to play, pause, or skipt the scene at the top of unity) every time it fires a projectile. I imagine it should be fixed once the Boss class is ready just a heads up.