A Lot of things, check description
updated boss prefab so it has the animations and the death splatter.
Moved fade-in and fade-out animations and controller to the animations folder.
Added text to display the level, and the score, and scripts to control them. Known bugs with the score is that it resets after every floor. To get scores for bugs and mages, copied the enemy_health script, renamed it enemy_health_bug and attached it to the bug. This way I could assign different values to the bugs and the mages when we kill them.